pyproject.toml
[tool.poetry]
name = "ioc-enrichment"
version = "0.1.0"
description = "ioc enrichment project"
authors = ["Buğra Çayır"]
readme = "README.md"
packages = [{include = "ioc_enrichment"}]

[tool.poetry.dependencies]
python = "^3.11"
celery = "^5.3.1"
redis = "^4.6.0"
sqlalchemy = "^2.0.19"
psycopg2 = "^2.9.6"
fastapi = "^0.100.1"
uvicorn = {extras = ["standard"], version = "^0.23.1"}
dynaconf = "^3.2.0"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
black = "^23.7.0"
bandit = "^1.7.5"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
